Refactor hide and block code
